Add build:clean script
a `clean` target is a common pattern with tradition makefile and I think is still useful here as removes junk and resets to a know condition..

I decided is simple so I did not make it a flavour script but it could be argued everything should be encapsulated so should be moved. I think eject works fine.